fix: split admin integration module
