+++ crates/dcc-mcp-gateway/src/gateway/admin/mod.rs
@@ -72,6 +72,8 @@ mod trace_log;
 mod traffic;
 #[cfg(feature = "admin")]
 mod update;
+#[cfg(feature = "admin")]
+mod wecom_url;
 pub mod workers;
 pub mod workflows;
